+++ src/com.tw.go.config.json/Capabilities.java
@@ -1,7 +1,14 @@
 package com.tw.go.config.json;
 
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
 public class Capabilities {
+    @Expose
+    @SerializedName("supports_pipeline_export")
     private boolean supportsPipelineExport;
+    @Expose
+    @SerializedName("supports_parse_content")
     private boolean supportsParseContent;
 
     public Capabilities() {